Made a minor tweak to urls.py after failing to fix certain variable names
Created the framework for the add/edit/deleteuser forms (Not correct, but a used the login view as a framework. Nick had some confusion on what should go in these views

handing off to the team to edit forms.py.

Mansi, Noah, and Nick all confirmed that the run_inv will not connect to the localhost. We believe this may be due to the work that we are doing with views and forms.